requirements.txt
# requirements.txt
# Prerequisites: Install PyTorch and FAISS via conda first (see install.sh)
# conda install pytorch==2.0.1 torchvision==0.15.2 torchaudio==2.0.2 pytorch-cuda=11.8 -c pytorch -c nvidia -y
# FOr windows, conda install -c pytorch faiss-cpu -y
# conda install -c pytorch faiss-cpu -y


timm==0.9.12

git+https://github.com/openai/CLIP.git
ftfy==6.1.1
regex==2023.8.8

numpy==1.24.3
pandas==2.0.3
Pillow==10.0.0
opencv-python-headless==4.8.0.74
scipy==1.11.1
albumentations==1.3.1

matplotlib==3.7.2
seaborn==0.12.2
tensorboard==2.13.0
wandb==0.15.5

scikit-learn==1.3.0

fastapi==0.100.0
uvicorn[standard]==0.23.1
python-multipart==0.0.6
aiofiles==23.1.0
jinja2==3.1.2

tqdm==4.65.0
pyyaml==6.0.1
h5py==3.9.0
gdown==4.7.1
einops==0.7.0

jupyter==1.0.0
ipython==8.14.0
pytest==7.4.0
black==23.7.0
flake8==6.1.0